requirements.txt
# Core AI and Language Processing
langchain==0.3.27
langchain-openai==0.3.33
openai==1.107.3

# Web Framework
fastapi==0.115.6
uvicorn==0.35.0

# Environment and Configuration
python-dotenv==1.1.1

# HTTP Requests
requests==2.32.5

# OCR and Image Processing
easyocr==1.7.2
opencv-python-headless==4.12.0.88
pillow==11.3.0
numpy==2.2.6

# Audio Processing
pyaudio==0.2.14
pygame==2.6.1
keyboard==0.13.5

# Email
# smtplib and email are built-in Python modules

# Additional dependencies (automatically installed with above packages)
# These are dependencies of the main packages but listed for clarity:
certifi==2025.8.3
charset-normalizer==3.4.3
httpx==0.28.1
pydantic==2.11.9
typing-extensions==4.15.0
urllib3==2.5.0